[android] - range alpha between 0 and 1 with android color conversion

+++ platform/android/MapboxGLAndroidSDK/src/main/java/com/mapbox/mapboxsdk/style/expressions/Expression.java
@@ -213,7 +213,7 @@ public class Expression {
    * @return the color expression
    */
   public static Expression color(@ColorInt int color) {
-    int[] rgba = PropertyFactory.colorToRgbaArray(color);
+    float[] rgba = PropertyFactory.colorToRgbaArray(color);
     return rgba(rgba[0], rgba[1], rgba[2], rgba[3]);
   }
